+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+## [1.8.2+1] 
+- Added locales generation complex mode
+
 ## [1.8.2]
 - Replace cli_dialog with dcli
 - version dependency updated
-- Added dafault value to menu & spaces
+- Added default value to menu & spaces
 - Fixed some lints with views
 
 ## [1.8.1]